Rename service.py to comicvine.py
Change cover_date to be Optional, closes #100
